neuro_prediction/requirements_nogpu.txt
alembic==1.13.1
asgiref==3.7.2
boto3==1.34.99
botocore==1.34.99
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
colorlog==6.8.0
contourpy==1.1.0
cycler==0.11.0
decorator==5.1.1
dill==0.3.7
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
filelock==3.9.0
fonttools==4.42.1
fsspec==2024.3.0
greenlet==3.0.3
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.5
lxml==4.9.3
Mako==1.3.0
Markdown==3.6
MarkupSafe==2.1.3
matplotlib==3.8.4
mne==1.4.0
mpmath==1.2.1
networkx==3.0
nibabel==5.1.0
nilearn==0.10.2
numpy==1.24.3
optuna==3.5.0
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
platformdirs==3.10.0
pooch==1.7.0
psycopg2-binary==2.9.9
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.23
sqlparse==0.4.4
sympy==1.11.1
threadpoolctl==3.2.0
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
